pyproject.toml
[project]
name = "obbb-research"
version = "1.0.0"
description = "One Big Beautiful Bill profit opportunity research database"
authors = [
    {name = "Your Name & Claude (Anthropic)"}
]
dependencies = [
    "beautifulsoup4>=4.12.2",
    "pandas>=2.1.4",
    "numpy>=1.24.3",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "openai>=1.3.0",
    "markdown>=3.5.1",
    "pyyaml>=6.0.1",
    "fast-graphrag>=0.0.5",
    "streamlit>=1.28.0",
    "plotly>=5.15.0",
    "jupyter>=1.0.0"
]
requires-python = ">=3.10"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]